backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-filter==23.2
django-import-export==3.2.0
django-modeladmin-reorder==0.3.1
djoser==2.2.0
drf-base64==2.0
Pillow==10.0.0
pip==23.2.1
python-dotenv==1.0.0
webcolors==1.13